[Bronze II] Title: 숫자, Time: 16 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-	int A, B;
+	long long A, B;
 	cin >> A >> B;
 
 	if (A == B || abs(A-B) == 1) {
@@ -12,14 +12,14 @@ int main() {
 	else if (A > B) {
 		cout << A - B - 1 << '\n';
 
-		for (int i = B + 1; i < A; i++) {
+		for (long long i = B + 1; i < A; i++) {
 			cout << i << " ";
 		}
 	}
 	else {
 		cout << B - A - 1 << '\n';
 
-		for (int i = A + 1; i < B; i++) {
+		for (long long i = A + 1; i < B; i++) {
 			cout << i << " ";
 		}
 	}
